+++ mobile/shared/components/geckoview/GeckoViewStartup.sys.mjs
@@ -225,6 +225,7 @@ export class GeckoViewStartup {
               "GeckoView:IPProtection:Deactivate",
               "GeckoView:IPProtection:Enroll",
               "GeckoView:IPProtection:RefreshUsage",
+              "GeckoView:IPProtection:ServerList:GetCountryList",
             ],
           });